UIEXPMGR-91: Refactor permissions to display link to download bulk-edit upload files
