+++ src/common/hooks/index.js
@@ -1,3 +1,4 @@
 export * from './useExportConfig';
 export * from './useExportJobScheduler';
 export * from './useSecureDownload';
+export * from './useExportManagerPerms';
